+++ lib/src/common/string_converter.dart
@@ -45,4 +45,29 @@ extension StringUtils on String {
       return defaultValue;
     }
   }
+
+  Map<String, List<String>> splitQuery({Encoding encoding = utf8}) {
+    return split("&").fold({}, (map, element) {
+      int index = element.indexOf("=");
+      if (index == -1) {
+        if (element != "") {
+          var s = Uri.decodeQueryComponent(element, encoding: encoding);
+          if (!map.containsKey(s)) {
+            map[s] = [];
+          }
+        }
+      } else if (index != 0) {
+        var key = element.substring(0, index);
+        var value = element.substring(index + 1);
+        var k = Uri.decodeQueryComponent(key, encoding: encoding);
+        var v = Uri.decodeQueryComponent(value, encoding: encoding);
+        if (map.containsKey(k)) {
+          map[k]!.add(v);
+        } else {
+          map[k] = [v];
+        }
+      }
+      return map;
+    });
+  }
 }