+++ lib/src/request/file_part.dart
@@ -1,0 +1,15 @@
+library cruky.request.file_part;
+
+class FilePart {
+  /// name of the field
+  final String name;
+
+  /// name of the file
+  final String filename;
+
+  /// streamed bytes of the files
+  final Stream<List<int>> bytes;
+
+  /// init
+  FilePart(this.name, this.filename, this.bytes);
+}